requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
Pillow==8.4.0
transformers==4.14.3
torch==1.10.0